+++ lib/countly_flutter.dart
@@ -2014,6 +2014,10 @@ class Countly {
         countlyConfig['autoEnrollABOnDownload'] = config.autoEnrollABOnDownload;
       }
 
+      if (config.requestDropAgeHours != null) {
+        countlyConfig['requestDropAgeHours'] = config.requestDropAgeHours;
+      }
+
       countlyConfig['remoteConfigAutomaticTriggers'] = config.remoteConfigAutomaticTriggers;
 
       countlyConfig['remoteConfigValueCaching'] = config.remoteConfigValueCaching;